Move the logic for ObjectLocation destinations into a separate trait
